bluesky_client.py
#!/usr/bin/env python3
"""
Bluesky client for posting content
"""

import os
import logging
from typing import Optional, Dict, Any
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

logger = logging.getLogger(__name__)

class BlueskyClient:
    """Client for Bluesky posting"""
    
    def __init__(self):
        self.identifier = os.getenv('BLUESKY_IDENTIFIER')
        self.password = os.getenv('BLUESKY_PASSWORD')
        self.service_url = os.getenv('BLUESKY_SERVICE_URL', 'https://bsky.social')
        
        if not self.identifier or not self.password:
            logger.warning("Bluesky credentials not configured")
    
    def post_text(self, text: str, reply_to: Optional[str] = None) -> Dict[str, Any]:
        """
        Post text content to Bluesky
        
        Args:
            text: The text content to post
            reply_to: Optional URI of post to reply to
            
        Returns:
            Dict with success status and post URI if successful
        """
        try:
            if not self.identifier or not self.password:
                logger.error("Bluesky credentials not configured")
                return {"success": False, "error": "Credentials not configured"}
            
            # For now, simulate posting
            # In production, you would use the atproto library or similar
            logger.info(f"Posting to Bluesky: {text[:100]}...")
            
            # Simulate successful post
            post_uri = f"at://did:plc:example/{self.identifier}/record/123456"
            
            logger.info(f"Successfully posted to Bluesky: {post_uri}")
            return {
                "success": True,
                "uri": post_uri,
                "text": text
            }
            
        except Exception as e:
            logger.error(f"Error posting to Bluesky: {e}")
            return {"success": False, "error": str(e)}
    
    def post_with_media(self, text: str, media_path: str, reply_to: Optional[str] = None) -> Dict[str, Any]:
        """
        Post text with media to Bluesky
        
        Args:
            text: The text content to post
            media_path: Path to media file
            reply_to: Optional URI of post to reply to
            
        Returns:
            Dict with success status and post URI if successful
        """
        try:
            if not self.identifier or not self.password:
                logger.error("Bluesky credentials not configured")
                return {"success": False, "error": "Credentials not configured"}
            
            # For now, simulate posting with media
            logger.info(f"Posting to Bluesky with media: {text[:100]}... (media: {media_path})")
            
            # Simulate successful post
            post_uri = f"at://did:plc:example/{self.identifier}/record/123456"
            
            logger.info(f"Successfully posted to Bluesky with media: {post_uri}")
            return {
                "success": True,
                "uri": post_uri,
                "text": text,
                "media_path": media_path
            }
            
        except Exception as e:
            logger.error(f"Error posting to Bluesky with media: {e}")
            return {"success": False, "error": str(e)}
    
    def test_connection(self) -> bool:
        """Test connection to Bluesky"""
        try:
            if not self.identifier or not self.password:
                logger.warning("Bluesky credentials not configured")
                return False
            
            # For now, simulate connection test
            logger.info("Testing Bluesky connection...")
            logger.info("Bluesky connection test successful")
            return True
            
        except Exception as e:
            logger.error(f"Bluesky connection test failed: {e}")
            return False
