requirements_moderation.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
jinja2>=3.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0
bcrypt>=4.0.0
requests>=2.31.0
psutil>=5.9.0
